基礎程式設計

這是2023年11月13日開始的，由陳建志老師教授的基礎程式語言課程(GA1002)資訊頁面。 請務必閱讀 課程須知。 若有任何問題，請參照助教與問題回報。

基礎程式設計

課程須知

測驗與計分方式

測驗時間、範圍與流程

測驗規定

附註

助教與問題回報

課堂測驗成績

成績分布

個人成績查詢

BasicProgrammingNCUES1121.DataHolder

BasicProgrammingNCUES1121.lazymaillist

BasicProgrammingNCUES1121.lazymaillist

BasicProgrammingNCUES1121.makewide!

BasicProgrammingNCUES1121.prosheet

BasicProgrammingNCUES1121.prosheet!

BasicProgrammingNCUES1121.readgsheet

BasicProgrammingNCUES1121.DataHolder — Type

mutable struct DataHolder
    data::DataFrame
    GSID::GoogleSheetIdentifier
end

BasicProgrammingNCUES1121.lazymaillist — Method

lazymaillist(path) read the CSV file's "Email" column, and returns a string of addresses separated by comma (for GMail) by default.

BasicProgrammingNCUES1121.lazymaillist — Method

lazymaillist() reads CSV in ./data/BasicProgrammingStudentList_112-1.csv. It will fail if the current directory is wrong or there is no such data (dvc pulled?).

BasicProgrammingNCUES1121.makewide! — Method

makewide!(dh::DataHolder)

BasicProgrammingNCUES1121.prosheet! — Method

prosheet!(dh::DataHolder)

BasicProgrammingNCUES1121.prosheet — Method

prosheet(df::DataFrame, RawScore) processes df = readgsheet(RawScore()). It returns DataFrame.

BasicProgrammingNCUES1121.readgsheet — Method

Given GSID<:GoogleSheetIdentifier, readgsheet(GSID::GoogleSheetIdentifier) returns a DataHolder storing the data obtained from the google sheet. GSID must contain keys to url for indexing into dir_local("credentials.json") to get the url. The url in dir_local("credentials.json") must be public.

Example

Read data on cloud in dir_local("credentials.json") at entry ["RawScore"]["url"].

readgsheet(RawScore())

This function depends on GoogleDrive.jl, it is designed to work with DVC to ensure secret data/path/url won't be push to public.

Noted that there is another package GoogleSheets.jl. However, this requires more complex pre settings:

Enable Google Sheets API:

Create a new project or select an existing one in Google Cloud Console.

Create OAuth client ID and enable the Google Sheets API for your project.

Create credentials for the API and download the JSON file, and follow the example of GoogleSheets.jl